implement d12 parts 1,2
part 1 required to move from initial position in the directions
of the instructions. F (forward) corresponded to moving in the
direction currently faced. Therefore instruction F is first converted
to the instruction of the direction currently faced, then instruction
is carried out as normal.
Part 2, however, features a waypoint. The waypoint is the only thing
being moved until F instruction occurs. F means move in the direction
of the waypoint a number of times equal to the instruction's number.